Initial refactor
