+++ bin/climb
@@ -7,7 +7,7 @@ if (file_exists(__DIR__.'/../vendor/autoload.php')) {
     require __DIR__.'/../../../autoload.php';
 }
 
-$app = new Vinkla\Climb\Application(
+$app = new Vinkla\Climb\Console\Application(
     realpath(__DIR__)
 );
