+++ composer.json
@@ -14,7 +14,6 @@
         "symfony/console": "^2.7.3",
         "league/climate": "^3.2.0",
         "knplabs/packagist-api": "^1.3.0",
-        "danielstjules/stringy": "^2.0.0",
         "composer/semver": "^1.0.0"
     },
     "require-dev": {
