+++ src/ClimbException.php
@@ -1,23 +1,9 @@
 <?php
 
-/*
- * This file is part of Climb.
- *
- * (c) Vincent Klaiber <hello@vinkla.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Vinkla\Climb;
+namespace Vinkla\Climb\Console;
 
 use Exception;
 
-/**
- * This is the climb exception class.
- *
- * @author Vincent Klaiber <hello@vinkla.com>
- */
 class ClimbException extends Exception
 {
     //
