+++ test.php
@@ -1,0 +1,7 @@
+<?php require 'vendor/autoload.php';
+
+use Vinkla\Climb\Ladder;
+
+$ladder = new Ladder();
+
+var_dump($ladder->getOutdatedPackages());
